requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.0
Django==3.2.6
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-phonenumber-field==5.2.0
future==0.18.2
geocoder==1.38.1
idna==3.3
phonenumbers==8.12.30
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
